Search implemented without filtering.

+++ CoreSmart/Entities/Properties.cs
@@ -14,7 +14,7 @@ namespace CoreSmart.Entities
         public string City { get; set; }
         public string Market { get; set; }
         public string State { get; set; }
-        public double Lat { get; set; }
-        public double Lng { get; set; }
+        public float Lat { get; set; }
+        public float Lng { get; set; }
     }
 }
