+++ SmartSearchExperiments/Models/Properties.cs
@@ -14,8 +14,8 @@ namespace SmartSearchExperiments.Models
         public string City { get; set; }
         public string Market { get; set; }
         public string State { get; set; }
-        public double Lat { get; set; }
-        public double Lng { get; set; }
+        public float Lat { get; set; }
+        public float Lng { get; set; }
 
     }
 }
